Rename the install:addon command to just install. Throw friendly errors for install:npm and uninstall:npm.
