+++ lib/commands/install-npm.js
@@ -1,28 +1,21 @@
 'use strict';
 
-var Command = require('../models/command');
+var Command     = require('../models/command');
+var SilentError = require('../errors/silent');
+var Promise     = require('../ext/promise');
 
 module.exports = Command.extend({
   name: 'install:npm',
-  description: 'Installs npm packages.',
+  description: 'Npm package installs are now managed by the user.',
   works: 'insideProject',
 
   anonymousOptions: [
     '<package-names...>'
   ],
 
-  run: function(commandOptions, rawArgs) {
-    var NpmInstallTask = this.tasks.NpmInstall;
-    var npmInstall     = new NpmInstallTask({
-      ui:         this.ui,
-      analytics:  this.analytics,
-      project:    this.project
-    });
-
-    return npmInstall.run({
-      packages: rawArgs,
-      'save-dev': true,
-      'save-exact': true
-    });
+  run: function() {
+    var err  = 'This command has been deprecated. Please use `npm install ';
+    err     += '<packageName> --save-dev --save-exact` instead.';
+    return Promise.reject(new SilentError(err));
   }
 });
