+++ lib/commands/install.js
@@ -1,36 +1,42 @@
 'use strict';
 
-var Command     = require('../models/command');
-var Promise     = require('../ext/promise');
+var Command = require('../models/command');
 var SilentError = require('../errors/silent');
+var Promise = require('../ext/promise');
 
 module.exports = Command.extend({
   name: 'install',
-  description: 'Installs npm and bower packages.',
+  description: 'Installs an ember-cli addon from npm.',
   works: 'insideProject',
 
+  anonymousOptions: [
+    '<addon-name>'
+  ],
+
   run: function(commandOptions, rawArgs) {
-    if (rawArgs.length) {
-      var err = new SilentError('The `install` command does not take any arguments. You must use `install:npm` or `install:bower` to install a specific package.');
-      return Promise.reject(err);
+    if (!rawArgs.length) {
+      var msg  = 'The `install` command must take an argument with the name';
+      msg     += ' of an ember-cli addon. For installing all npm and bower ';
+      msg     += 'dependencies you can run `npm install && bower install`.';
+      return Promise.reject(new SilentError(msg));
     }
 
-    var BowerInstallTask = this.tasks.BowerInstall;
-    var bowerInstall     = new BowerInstallTask({
-      ui:         this.ui,
-      analytics:  this.analytics,
-      project:    this.project
+    var AddonInstallTask = this.tasks.AddonInstall;
+    var addonInstall = new AddonInstallTask({
+      ui:              this.ui,
+      analytics:       this.analytics,
+      project:         this.project,
+      NpmInstallTask:  this.tasks.NpmInstall,
+      BlueprintTask:   this.tasks.GenerateFromBlueprint
     });
 
-    var NpmInstallTask = this.tasks.NpmInstall;
-    var npmInstall     = new NpmInstallTask({
-      ui:         this.ui,
-      analytics:  this.analytics,
-      project:    this.project
-    });
+    var packageName = rawArgs.slice(0,1)[0];
+    var extraArgs   = rawArgs.slice(1);
 
-    return bowerInstall.run({}).then(function() {
-      return npmInstall.run({});
+    return addonInstall.run({
+      'package':         packageName,
+      extraArgs:         extraArgs,
+      blueprintOptions:  commandOptions
     });
   }
 });
