+++ lib/commands/uninstall-npm.js
@@ -1,28 +1,21 @@
 'use strict';
 
-var Command = require('../models/command');
+var Command     = require('../models/command');
+var SilentError = require('../errors/silent');
+var Promise     = require('../ext/promise');
 
 module.exports = Command.extend({
   name: 'uninstall:npm',
-  description: 'Uninstalls npm packages.',
+  description: 'Npm package uninstall are now managed by the user.',
   works: 'insideProject',
 
   anonymousOptions: [
     '<package-names...>'
   ],
 
-  run: function(commandOptions, rawArgs) {
-    var NpmUninstallTask = this.tasks.NpmUninstall;
-    var npmUninstall     = new NpmUninstallTask({
-      ui:         this.ui,
-      analytics:  this.analytics,
-      project:    this.project
-    });
-
-    return npmUninstall.run({
-      packages: rawArgs,
-      'save-dev': true,
-      'save-exact': true
-    });
+  run: function() {
+    var err  = 'This command has been deprecated. Please use `npm uninstall ';
+    err     += '<packageName> --save-dev` instead.';
+    return Promise.reject(new SilentError(err));
   }
 });
